UBUNTU: [Config] include include/generated/compile.h
When we prepare the headers all the generated header files are
built and installed into the new headers directory, all except for
include/generated/compile.h which is actually only made when version.o
is built.  This header is used and needed by several external modules
for their version checks.

As we build the headers after we build the main build directory we are
able to pull this header out of the binary build directory.  All other
generated headers are build during the prepare phase.

BugLink: http://bugs.launchpad.net/bugs/942569
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>